Modify version equality to allow comparison with strings
Updated the Version#== method to covert self to a string and then do a
comparison with the other string. I thought this would help make the
interface more natural.

The specs compare versions to strings a lot, and manually converting to
string each time felt tedious and interrupted the flow of the
assertions.